+++ R/utils.R
@@ -41,7 +41,7 @@ update.opsr <- function(object, ...) {
   NextMethod("update", object)
 }
 
-## returns loglik of ordinal probit model
+## returns loglik of ordered probit model
 ll_probit <- function(object) {
   probs <- lapply(seq_len(object$nReg), function(i) {
     predict(object, group = i, type = "prob")
@@ -53,14 +53,15 @@ ll_probit <- function(object) {
 r2 <- function(object) {
   z <- get_z(object)
   y <- get_y(object)
-  RS <- residuals(object)**2
-  TS <- (y - mean(y))**2
+  RS <- residuals(object)^2
+  TS <- (y - mean(y))^2
   R2o <- unlist(lapply(seq_len(object$nReg), function(i) {
     RSS <- sum(RS[z == i])
-    TSS <- sum(TS[z == i])
-    1 - RSS / TSS
+    yo <- y[z == i]
+    TSS <- sum((yo - mean(yo))^2)
+    1 - RSS/TSS
   }))
-  R2total <- 1 - sum(RS) / sum(TS)
+  R2total <- 1 - sum(RS)/sum(TS)
   R2 <- c(R2total, R2o)
   names(R2) <- c("Total", paste0("o", 1:object$nReg))
   R2
@@ -89,3 +90,23 @@ is_opsr_null <- function(object) {
 `%||%` <- function(x, y) {
   if (is.null(x)) y else x
 }
+
+## creds to stats::printCoefmat
+to.signif.codes <- function(pv) {
+  Signif <- stats::symnum(pv, corr = FALSE, na = FALSE,
+                          cutpoints = c(0, 0.001, 0.01, 0.05, 0.1, 1),
+                          symbols = c("***", "**", "*", ".", " "))
+  Signif
+}
+
+pleaseCite <- function(pkgname) {
+  pc <- utils::capture.output(print(utils::citation(pkgname)))
+  pc <- paste(pc, collapse = "\n")
+  pc
+}
+
+is_tobit_5 <- function(object) {
+  methods::is(object, "tobit.5") || object$nReg == 2
+}
+
+